Word and Address -> usize
Works the same on x86_64, more platform independent although nothing else is.

+++ tests/read-and-write.rs
@@ -22,7 +22,7 @@ fn test_attach_detach() {
 fn test_read() {
     let (buf_addr, pid) = fork_with_buffer("foobar");
     let reader = ptrace::Reader::new(pid);
-    match reader.peek_data(unsafe { buf_addr.offset(3) } as u64) {
+    match reader.peek_data(unsafe { buf_addr.offset(3) } as usize) {
         Ok(v) => assert_eq!((v & 0xff) as u8, 'b' as u8),
         Err(_) => panic!("Error while reading: {:?}", os_errno())
     }
@@ -32,7 +32,7 @@ fn test_read() {
 fn test_read_string() {
     let (buf_addr, pid) = fork_with_buffer("foobar");
     let reader = ptrace::Reader::new(pid);
-    match reader.read_string(buf_addr as u64) {
+    match reader.read_string(buf_addr as usize) {
         Ok(v) =>
             assert_eq!(v, vec!('f' as u8, 'o' as u8, 'o' as u8, 'b' as u8, 'a' as u8, 'r' as u8)),
         Err(_) =>
@@ -46,10 +46,10 @@ fn test_write() {
     let foo_word = 0x0123456789abcdef;
 
     let writer = ptrace::Writer::new(pid);
-    match writer.poke_data(buf_addr as u64, foo_word) {
+    match writer.poke_data(buf_addr as usize, foo_word) {
         Ok(_) => {
             let reader = ptrace::Reader::new(pid);
-            let v = reader.peek_data(buf_addr as u64).ok().expect("Could not read back word");
+            let v = reader.peek_data(buf_addr as usize).ok().expect("Could not read back word");
             assert_eq!(v, foo_word);
         },
         Err(_) =>
@@ -64,10 +64,10 @@ fn test_write_small_buf() {
     let (buf_addr, pid) = fork_with_buffer("foobar and then some");
     let writer = ptrace::Writer::new(pid);
     let buf = vec!('F' as u8, 'O' as u8, 'O' as u8, 'B' as u8, 'A' as u8, 'R' as u8);
-    match writer.write_data(buf_addr as u64, &buf) {
+    match writer.write_data(buf_addr as usize, &buf) {
         Ok(_) => {
             let reader = ptrace::Reader::new(pid);
-            let v = reader.read_string(buf_addr as u64).ok().expect("Could not read back buffer");
+            let v = reader.read_string(buf_addr as usize).ok().expect("Could not read back buffer");
             assert_eq!(str::from_utf8(v.as_slice()), Ok("FOOBAR and then some"));
         },
         Err(_) =>
@@ -84,10 +84,10 @@ fn test_write_large_buf() {
     let writer = ptrace::Writer::new(pid);
     let mut buf: Vec<u8> = Vec::new();
     buf.extend("FRIDDLE FRITZ FROB BAZ BAR FOO".as_bytes());
-    match writer.write_data(buf_addr as u64, &buf) {
+    match writer.write_data(buf_addr as usize, &buf) {
         Ok(_) => {
             let reader = ptrace::Reader::new(pid);
-            let v = reader.read_string(buf_addr as u64).ok().expect("Could not read back buffer");
+            let v = reader.read_string(buf_addr as usize).ok().expect("Could not read back buffer");
             assert_eq!(str::from_utf8(v.as_slice()), str::from_utf8(buf.as_slice()));
         },
         Err(_) =>
